+++ src/components/SummaryTable/index.jsx
@@ -3,11 +3,13 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import StatusWidget from '../StatusWidget';
 import CONFIG from '../../utils/nimbledroid/config';
-import { frum } from '../../vendor/queryOps';
+import { selectFrom } from '../../vendor/vectors';
 
 const SummaryTable = ({ content = [], header }) => {
   const compareName = CONFIG.packageIdLabels[CONFIG.baseProduct];
-  const compareColumn = frum(header).findIndex(name => name === compareName);
+  const compareColumn = selectFrom(header).findIndex(
+    name => name === compareName
+  );
 
   return (
     <table className="summary-table">
